Ordered hash reading compatibility
This commit changes the code to generate non-ordered hashes but be able
to read them. This allows for ordered hashes to be released in two
stages:

 1. Add ability to read ordered hashes, write unordered hashes
 2. Write ordered hashes

Therefore, there isn't a period of time where one part of the system or
cluster is generating information another part of the system cannot
read.